+++ CHANGELOG.md
@@ -3106,3 +3106,4 @@
 - [2026-07-24 10:18] test(network): implement widget integration tests for network
 - [2026-07-24 12:12] perf(test): apply RepaintBoundary on animated test widgets
 - [2026-07-24 12:51] feat(ui): implement responsive ui view component
+- [2026-07-24 13:39] feat(auth): implement responsive auth view component